Now works very well (Fixed massive bug)
This version works well and fast to evolve very good food-chasers. This
is because I fixed a massive bug in reproduction.

Self-sustaining populations at ~12 agents.

As expected, evolutionary progress does not result in more agents, but
in lower density of food. Relationship with cosmic-scale criteria of
evolutionary progress as energy capture / Kardashev scale?

(Food energy is quite high at 0.5, and 25% of the reproductions are
identical without mutation; this might change)