Perhaps some deep diversity in large world?
java World VISUAL 1 SEED 5 FIGHTENERGY .03 EATBONUS 100000.0 FIGHTDAMAGE 4.0 SPEEDENERGY .02 ENERGYDECAY .995 WSIZE 3005 MEANADDEDFOODPERSTEP 7
gets persistent diversity, with one lineage coalescing 150+
lifetimes!

And several more coalescing within 40-80 lifetimes...

But the diversity is not easy to see - seems mostly sessile, with
apparently a few high-aggro high-speed darting through, rarely... Note
that only a small portion of the space can be seen... but still...
Need to be able to save the population, and show half-sized workld for
visual!

Also, not clear if the deep diversity will persist... seems to be
decreasing!