W/ validneur, show diversity, not sur if boring...
Now using the valid-neur mask, so evolution can decide which neurons to
use...

The diversity is durable. It seems a bit boring: sessiles vs. hunters,
and the hunters don't even seem to have any turning ability!

BUT... some of the "Sessiles"  actually become mobile + high-aggro when
a hunter touches them!